Move molecular database import code to engine
* fix migration: reset molecular_db private key seq
* disable mol_db service deployment in Ansible
